feat: file modifications fixed, file operations moved to mixin
