+++ lib/app/util/extenstion/string_extensions.dart
@@ -1,0 +1,9 @@
+extension ClearString on String {
+  String clearSpaces() {
+    var result = this;
+    while (result.contains('  ')) {
+      result = result.replaceAll('  ', ' ');
+    }
+    return result;
+  }
+}
