+++ lib/domain/service/component_generator/file_operation_result.dart
@@ -1,0 +1,6 @@
+enum FileOperationResult {
+  created,
+  modified,
+  alreadyExists,
+  error,
+}
